requirements.txt
numpy
pandas
matplotlib
scipy
scikit-learn
keras
opencv-python
torchvision
ultralytics
albumentations
sqlite3